# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FLOWLINE Crowd API",
    "description": "Crowd density, bottleneck detection, and rerouting for venue operations. Inference via Hugging Face rootstrap-org/crowd-counting (CSRNet).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/demo/tick": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Get Demo Tick",
        "operationId": "get_demo_tick",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 120,
              "minimum": 0,
              "default": 0,
              "title": "T"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analyze": {
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "Analyze",
        "operationId": "analyze",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ops-analyze"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_ops-analyze": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "reset": {
            "type": "boolean",
            "title": "Reset",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ops-analyze"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelMeta": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "model_repo": {
            "type": "string",
            "title": "Model Repo"
          },
          "global_count": {
            "type": "number",
            "title": "Global Count"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "source",
          "model_repo",
          "global_count",
          "ready"
        ],
        "title": "ModelMeta"
      },
...
      "RouteSuggestion": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From"
          },
          "to": {
            "type": "string",
            "title": "To"
          },
          "avoid_zones": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avoid Zones"
          },
          "path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Path"
          },
          "path_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Path Labels"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "path",
          "path_labels",
          "message"
        ],
        "title": "RouteSuggestion"
      },
      "Snapshot": {
        "properties": {
          "venue_id": {
            "type": "string",
            "title": "Venue Id"
          },
          "venue_name": {
            "type": "string",
            "title": "Venue Name"
          },
          "t": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "T"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ZoneSnapshot"
            },
            "type": "array",
            "title": "Zones"
          },
          "bottlenecks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bottlenecks"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/RouteSuggestion"
            },
            "type": "array",
            "title": "Routes"
          },
          "suggestion": {
            "type": "string",
            "title": "Suggestion"
          },
          "model": {
            "$ref": "#/components/schemas/ModelMeta"
          },
          "history": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "object",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "venue_id",
          "venue_name",
          "zones",
          "bottlenecks",
          "routes",
          "suggestion",
          "model",
          "history"
        ],
        "title": "Snapshot"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ZoneSnapshot": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "capacity": {
            "type": "number",
            "title": "Capacity"
          },
          "count_est": {
            "type": "number",
            "title": "Count Est"
          },
          "density": {
            "type": "number",
            "title": "Density"
          },
          "risk": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Risk"
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid"
          },
          "polygon": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Polygon"
          },
          "trend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type",
          "capacity",
          "count_est",
          "density",
          "risk",
          "centroid",
          "polygon"
        ],
        "title": "ZoneSnapshot"
      }
    }
  }
}